requirements.txt
numpy
scipy
h5py
torch
torchvision
torchaudio
timm
matplotlib
seaborn
einops
thop
pyyaml
loris
tonic
dv